# rest_api/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for TON API interaction written on Golang using \"Stonfi\" API, SDK \"tonapi-go\" and SDK \"tongo\".\nAll resources is protected with api-key in query.",
        "title": "RESTful API for TON API interaction",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "paths": {
        "/account/get-jetton": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get jetton balance on account and other info about jetton by it master address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get jetton balance on account",
                "operationId": "get-jetton",
                "parameters": [
                    {
                        "type": "string",
                        "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O",
                        "description": "мастер-адрес монеты (jetton_master)",
                        "name": "masterAddress",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.AccountJetton"
                        }
                    }
                }
            }
        },
        "/account/get-jettons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all non-null jettons balances on account and other info about jettons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get jettons balances on account",
                "operationId": "get-jettons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/account.AccountJetton"
                            }
                        }
                    }
                }
            }
        },
        "/account/get-seqno": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get account seqno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get account seqno",
                "operationId": "get-seqno",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.GetSeqnoOut"
                        }
                    }
                }
            }
        },
        "/account/get-ton": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get TON balance on account and other info about TON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get TON balance on account",
                "operationId": "get-ton",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/account.TonJetton"
                            }
                        }
                    }
                }
            }
        },
        "/jettons/get-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get jetton info from Stonfi API by it master address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jettons"
                ],
                "summary": "Get jetton info",
                "operationId": "get-info",
                "parameters": [
                    {
                        "type": "string",
                        "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O",
                        "description": "мастер-адрес монеты (jetton_master)",
                        "name": "masterAddress",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jettons.JettonParams"
                        }
                    }
                }
            }
        },
        "/transactions/buy/pre-request": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get pre-request info about buy transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Buy pre-request",
                "operationId": "buy-pre-request",
                "parameters": [
                    {
                        "type": "number",
                        "example": 0.1,
                        "description": "кол-во используемых TON для покупки в формате, удобном для человека",
                        "name": "amount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O",
                        "description": "мастер-адрес покупаемой монеты (jetton_master)",
                        "name": "jettonCA",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "процент проскальзывания",
                        "name": "slippage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.PreRequestBuyJetton"
                        }
                    }
                }
            }
        },
        "/transactions/buy/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send transaction to buy jettons using TON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Buy send",
                "operationId": "buy-send",
                "parameters": [
                    {
                        "description": "BuySendIn struct params",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.BuySendIn"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializers.BuySendOut"
                        }
                    }
                }
            }
        },
        "/transactions/cell/pre-request": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get pre-request info about cell transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Cell pre-request",
                "operationId": "cell-pre-request",
                "parameters": [
                    {
                        "type": "number",
                        "example": 200,
                        "description": "кол-во используемых монет на продажу в формате, удобном для человека",
                        "name": "amount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O",
                        "description": "мастер-адрес продаваемой монеты (jetton_master)",
                        "name": "jettonCA",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "процент проскальзывания",
                        "name": "slippage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.PreRequestCellJetton"
                        }
                    }
                }
            }
        },
        "/transactions/cell/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send transaction to cell jettons to TON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Cell send",
                "operationId": "cell-send",
                "parameters": [
                    {
                        "description": "CellSendIn struct params",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.CellSendIn"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializers.CellSendOut"
                        }
                    }
                }
            }
        },
        "/transactions/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get transaction info by given its hash and action (buy OR cell)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Transaction info",
                "operationId": "transactions/-info",
                "parameters": [
                    {
                        "type": "string",
                        "example": "cell",
                        "description": "действие с монетами в транзакции (покупка/продажа)",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "29a301e4d2a05713f4eab6c8f0daa3c58eed15d1d41678068cd50fe46ca7f6a5",
                        "description": "хэш транзакции",
                        "name": "transactionHash",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.TransactionInfoWithStatusOK"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "account.AccountJetton": {
            "description": "Описывает монету (кроме TON), имеющуюся у аккаунта",
            "type": "object",
            "properties": {
                "balance": {
                    "description": "баланс монеты на аккаунте",
                    "type": "integer",
                    "example": 326166742480
                },
                "beautyBalance": {
                    "description": "округлённый баланс",
                    "type": "string",
                    "example": "326.167"
                },
                "decimals": {
                    "description": "decimals монеты",
                    "type": "integer",
                    "example": 9
                },
                "masterAddress": {
                    "description": "мастер-адрес монеты (jetton_master)",
                    "type": "string",
                    "example": "0:b8ef4f77a17e5785bd31ba4da50abd91852f2b8febee97ad6ee16d941f939198"
                },
                "symbol": {
                    "description": "символ монеты",
                    "type": "string",
                    "example": "GRAM"
                }
            }
        },
        "account.TonJetton": {
            "description": "Описывает TON монету, имеющуюся у аккаунта",
            "type": "object",
            "properties": {
                "balance": {
                    "description": "баланс монеты на аккаунте",
                    "type": "integer",
                    "example": 1955532900
                },
                "beautyBalance": {
                    "description": "округлённый баланс",
                    "type": "string",
                    "example": "1.956"
                },
                "decimals": {
                    "description": "decimals монеты",
                    "type": "integer",
                    "example": 9
                }
            }
        },
        "jettons.JettonParams": {
            "description": "описывает любую монету (функция получения информации о монете по её адресу)",
            "type": "object",
            "properties": {
                "decimals": {
                    "description": "decimals монеты",
                    "type": "integer",
                    "example": 9
                },
                "masterAddress": {
                    "description": "мастер-адрес монеты (jetton_master)",
                    "type": "string",
                    "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O"
                },
                "priceUsd": {
                    "description": "цена монеты в долларах (USD)",
                    "type": "number",
                    "example": 0.002695039717551585
                },
                "symbol": {
                    "description": "символ монеты",
                    "type": "string",
                    "example": "GRAM"
                }
            }
        },
        "serializers.BuySendIn": {
            "description": "Cтруктура входных данных для отправки транзакции на покупку",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "кол-во используемых TON для покупки в формате, удобном для человека",
                    "type": "number",
                    "example": 0.1
                },
                "jettonCA": {
                    "description": "мастер-адрес покупаемой монеты (jetton_master)",
                    "type": "string",
                    "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O"
                },
                "slippage": {
                    "description": "процент проскальзывания",
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "serializers.BuySendOut": {
            "description": "Успешная отправка транзакции на покупку",
            "type": "object",
            "properties": {
                "success": {
                    "description": "успех",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "serializers.CellSendIn": {
            "description": "Cтруктура входных данных для отправки транзакции на продажу",
            "type": "object",
            "properties": {
                "amount": {
                    "description": "кол-во используемых монет на продажу в формате, удобном для человека",
                    "type": "number",
                    "example": 200
                },
                "jettonCA": {
                    "description": "мастер-адрес продаваемой монеты (jetton_master)",
                    "type": "string",
                    "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O"
                },
                "slippage": {
                    "description": "процент проскальзывания",
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "serializers.CellSendOut": {
            "description": "Успешная отправка транзакции на продажу",
            "type": "object",
            "properties": {
                "success": {
                    "description": "успех",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "serializers.GetSeqnoOut": {
            "description": "Seqno аккаунта",
            "type": "object",
            "properties": {
                "seqno": {
                    "description": "порядковый номер версии кошелька аккаунта",
                    "type": "string",
                    "example": "105"
                }
            }
        },
        "transactions.PreRequestBuyJetton": {
            "type": "object",
            "properties": {
                "dex": {
                    "description": "название DEX биржи",
                    "type": "string",
                    "example": "Stonfi"
                },
                "jettonCA": {
                    "description": "мастер-адрес покупаемой монеты (jetton_master)",
                    "type": "string",
                    "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O"
                },
                "jettonSymbol": {
                    "description": "символ покупаемой монеты",
                    "type": "string",
                    "example": "GRAM"
                },
                "jettonsOut": {
                    "description": "примерное кол-во монет, которые будут получены после транзакции",
                    "type": "string",
                    "example": "175.137"
                },
                "minOut": {
                    "description": "минимальное кол-во покупаемых монет (с учётом процента проскальзывания)",
                    "type": "string",
                    "example": "140.11"
                },
                "usedTon": {
                    "description": "кол-во используемых TON для покупки в формате, удобном для человека",
                    "type": "string",
                    "example": "0.1"
                }
            }
        },
        "transactions.PreRequestCellJetton": {
            "type": "object",
            "properties": {
                "dex": {
                    "description": "название DEX биржи",
                    "type": "string",
                    "example": "Stonfi"
                },
                "jettonCA": {
                    "description": "мастер-адрес продаваемой монеты (jetton_master)",
                    "type": "string",
                    "example": "EQC47093oX5Xhb0xuk2lCr2RhS8rj-vul61u4W2UH5ORmG_O"
                },
                "jettonSymbol": {
                    "description": "символ продаваемой монеты",
                    "type": "string",
                    "example": "GRAM"
                },
                "minOut": {
                    "description": "минимальное кол-во получаемых монет TON (с учётом процента проскальзывания)",
                    "type": "string",
                    "example": "0.091"
                },
                "tonsOut": {
                    "description": "примерное кол-во монет TON, которые будут получены после транзакции",
                    "type": "string",
                    "example": "0.114"
                },
                "usedJettons": {
                    "description": "кол-во используемых монет на продажу в формате, удобном для человека",
                    "type": "string",
                    "example": "200.0"
                }
            }
        },
        "transactions.TransactionInfoWithStatusOK": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "действие с монетами в транзакции (покупка/продажа)",
                    "type": "string",
                    "example": "buy"
                },
                "bounce": {
                    "description": "была ли отклонена транзакция (не означает успех или неудачу транзакции)",
                    "type": "boolean",
                    "example": true
                },
                "endBalance": {
                    "description": "оставшееся кол-во TON после проведения транзакции",
                    "type": "string",
                    "example": "2.689"
                },
                "hash": {
                    "description": "хэш транзакции",
                    "type": "string",
                    "example": "ed79dafdda1a766dc6d7745e8dd03410adf7ba57bb6fccdb33ee5f5d8c3640f4"
                },
                "opName": {
                    "description": "название операции транзакции",
                    "type": "string",
                    "example": "jetton_notify"
                },
                "statusOK": {
                    "description": "успех или неудача выполнения транзакции",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Security api key. Please add it to URL like \"?api-key=5how45gi54yi\" to authorize your requests.",
            "type": "apiKey",
            "name": "api-key",
            "in": "query"
        }
    }
}